Work on mock client for tests
Okay, so you're probably like "you put all this work on the tests and almost
nothing at all on the actual server code. Why?" The reason is that this is the
first implementation of the AC network protocol, and so I am establishing the
correct idioms for writing a client, since the mock client code will essentially
be ported over to the Python client once this is all done.

JSON schemas for most of the server -> client messages have been written. This
should probably be moved to its own repository and added back as a submodule,
so that the JS server and Python client protocols remain in sync, since JSON
schemas can also be used in Python for input validation.

I have also added an ESLint configuration, although it is not in the dev
dependencies of the package.json. You are welcome to install it yourself.

Coverage is expected to be abysmal for the foreseeable future due to branches
for input validation.